Reorganize sample service for kmt.py
