+++ pkg/dynamicinstrumentation/testutil/sample/.gitignore
