+++ pkg/dynamicinstrumentation/testutil/sample/strings.go
@@ -3,12 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package main
+package sample
 
 //nolint:all
 //go:noinline
 func test_single_string(x string) {}
 
-func executeStringFuncs() {
+//nolint:all
+func ExecuteStringFuncs() {
 	test_single_string("abc")
 }
